build demo page
