+++ docs/demo.html
@@ -10,7 +10,7 @@
     <script src="js/holder.js"></script>
     <script src="js/diff2html.min.js"></script>
     <script src="js/bootstrap.min.js"></script>
-    <script src="js/view.js"></script>
+    <script src="js/demo.bundle.js"></script>
 </head>
 
 <body>
