+++ package.json
@@ -7,13 +7,13 @@
   "scripts": {
     "test": "jest",
     "test:report": "jest --config jest-stare.config.js",
-    "build": "tsc -p ."
+    "build:demo": "tsc -p . && browserify lib/demo.js -o docs/js/demo.bundle.js"
   },
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "mathjs": "^9.3.0",
-    "lodash": "^4.17.21"
+    "lodash": "^4.17.21",
+    "mathjs": "^9.3.0"
   },
   "devDependencies": {
     "@babel/runtime": "^7.13.10",
@@ -22,19 +22,20 @@
     "@types/glob": "^7.1.3",
     "@types/jest": "^26.0.14",
     "@types/jest-diff": "^24.3.0",
-    "@types/minimist": "^1.2.0",
     "@types/mathjs": "^6.0.11",
+    "@types/minimist": "^1.2.0",
+    "browserify": "^17.0.0",
     "debug": "^4.3.1",
     "deep-diff": "^1.0.2",
     "fs-extra": "^9.0.1",
+    "glob": "^7.1.6",
     "jest": "^26.4.2",
     "jest-diff": "^26.4.2",
     "jest-silent-reporter": "^0.2.1",
     "jest-stare": "^2.2.1",
     "katex": "^0.13.0",
-    "glob": "^7.1.6",
-    "latex.js": "^0.12.4",
     "latex-utensils": "^2.0.3",
+    "latex.js": "^0.12.4",
     "minimist": "^1.2.5",
     "ts-jest": "^26.3.0",
     "ts-node": "^9.0.0",
